myapp/app/requirements.txt
# app/requirements.txt
fastapi
uvicorn
python-multipart
watchdog
jinja2
PyYAML